# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Abrpardaz API",
        "description": "This is the official API which can be used by any user",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1",
            "description": "Local server"
        },
        {
            "url": "http://abrpardaz.com/api/v1",
            "description": "The API endpoint"
        }
    ],
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "description": "",
                "operationId": "App\\Http\\Controllers\\V1\\AuthController::register",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login a new user",
                "description": "",
                "operationId": "App\\Http\\Controllers\\V1\\AuthController::login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/auth/forgetPassword": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Send a password reset link to the email",
                "description": "",
                "operationId": "App\\Http\\Controllers\\V1\\AuthController::forgetPassword",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/auth/resetPassword": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Changes the password using the provided token",
                "description": "",
                "operationId": "App\\Http\\Controllers\\V1\\AuthController::resetPassword",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/auth/changePassword": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Changes the user password",
                "description": "",
                "operationId": "App\\Http\\Controllers\\V1\\AuthController::changePassword",
                "parameters": [
                    {
                        "name": "current_password",
                        "in": "query",
                        "description": "The current password of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "new_password",
                        "in": "query",
                        "description": "The new password you want to set for the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "new_password_confirmation",
                        "in": "query",
                        "description": "Confirmation of the new password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/auth/verify": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Verify the user's email",
                "description": "",
                "operationId": "App\\Http\\Controllers\\V1\\AuthController::verify",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The email which user has been registered with",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "The token which has been sent to the user's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login and copy the token into the below box",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}